[Java] Continue renaming for streams, channels and connections, plus finish up keeping windowSize as an int.

+++ aeron-driver/src/main/java/uk/co/real_logic/aeron/driver/DefaultMulticastSenderControlStrategy.java
@@ -36,7 +36,7 @@ public class DefaultMulticastSenderControlStrategy implements SenderControlStrat
 
     /** {@inheritDoc} */
     public long onStatusMessage(final int termId, final long highestContiguousSequenceNumber,
-                                final long receiverWindowSize, final InetSocketAddress address)
+                                final int receiverWindowSize, final InetSocketAddress address)
     {
         final long positionLimit = ((long)termId << shiftsForTermId) + receiverWindowSize;
         this.positionLimit = Math.max(this.positionLimit, positionLimit);
